+++ Source/IGListAdapter.h
@@ -74,6 +74,11 @@ IGLK_SUBCLASSING_RESTRICTED
 @property (nonatomic, nullable, weak) id <UIScrollViewDelegate> scrollViewDelegate;
 
 /**
+ The updater for this list adapter.
+ */
+@property (nonatomic, strong, readonly) id <IGListUpdatingDelegate> updater;
+
+/**
  A bitmask of experiments to conduct on the adapter.
  */
 @property (nonatomic, assign) IGListExperiment experiments;
@@ -81,7 +86,7 @@ IGLK_SUBCLASSING_RESTRICTED
 /**
  Initializes a new `IGListAdapter` object.
 
- @param updatingDelegate An object that manages updates to the collection view.
+ @param updater An object that manages updates to the collection view.
  @param viewController   The view controller that will house the adapter.
  @param workingRangeSize The number of objects before and after the viewport to consider within the working range.
 
@@ -94,7 +99,7 @@ IGLK_SUBCLASSING_RESTRICTED
 
  To opt out of using the working range, you can provide a value of `0`.
  */
-- (instancetype)initWithUpdater:(id <IGListUpdatingDelegate>)updatingDelegate
+- (instancetype)initWithUpdater:(id <IGListUpdatingDelegate>)updater
                  viewController:(nullable UIViewController *)viewController
                workingRangeSize:(NSInteger)workingRangeSize NS_DESIGNATED_INITIALIZER;